Generic changes, refactoring, passing the success from db to the caller

+++ src/scenes.ts
@@ -6,12 +6,10 @@ interface MyWizardContext extends Scenes.WizardContext {
     wizard: any;
 }
 
-const cheersReply = async (ctx: BotContext) => {
+const cheersReply = async (ctx: BotContext, kmRounded: number) => {
     const stats = await db.getStatsForUser(ctx.from.id);
-
     return `
-    
-    Se oli kunnon repäsy.
+    Lisäsin sinulle ${kmRounded} kilometriä.
     
 Hyvä homma ${
         ctx.message.from.first_name
@@ -27,24 +25,21 @@ export const skiRecordWizard = new Scenes.WizardScene<MyWizardContext>(
         const reply = await ctx.reply('Ok, laitappas vielä ne kilometrit', {
             reply_markup: {
                 input_field_placeholder: '12.3',
-                // force the user to reply to the bot
                 force_reply: true,
                 one_time_keyboard: true,
             },
         });
 
         replyMessageId = reply.message_id;
-
         return ctx.wizard.next();
     },
     async (ctx) => {
-        // guard to check if the chat is valid
         if (!('text' in ctx.message)) {
             ctx.reply('Vastaa nyt järkevästi');
             return ctx.scene.reenter();
         }
+
         const kilometers = parseFloat(ctx.message.text.replace(',', '.'));
-        // smooth native rounding
         const kmRounded = Math.round(kilometers * 100) / 100;
 
         if (isNaN(kmRounded)) {
@@ -55,7 +50,7 @@ export const skiRecordWizard = new Scenes.WizardScene<MyWizardContext>(
             });
             return ctx.scene.reenter();
         } else {
-            await db.writeRecordToDb(
+            const result = await db.writeRecordToDb(
                 ctx.message.from.id,
                 ctx.message.from.first_name,
                 ctx.message.from.last_name,
@@ -63,20 +58,21 @@ export const skiRecordWizard = new Scenes.WizardScene<MyWizardContext>(
                 kmRounded,
             );
 
-            await ctx.reply(await cheersReply(ctx), {
-                reply_markup: {
-                    remove_keyboard: true,
-                },
-            });
+            if (result.success) {
+                await ctx.reply(await cheersReply(ctx, kmRounded), {
+                    reply_markup: {
+                        remove_keyboard: true,
+                    },
+                });
 
-            // wait for 1 second
-            await new Promise((resolve) => setTimeout(resolve, 1000));
+                await new Promise((resolve) => setTimeout(resolve, 1000));
 
-            // bot's message
-            await ctx.deleteMessage(replyMessageId);
-
-            // user's reply
-            await ctx.deleteMessage(ctx.message.message_id);
+                await ctx.deleteMessage(replyMessageId);
+                await ctx.deleteMessage(ctx.message.message_id);
+            } else {
+                ctx.reply('Jokin meni pieleen, yritä uudelleen.');
+                return ctx.scene.reenter();
+            }
         }
         return ctx.scene.leave();
     },
@@ -88,19 +84,17 @@ export const betWizard = new Scenes.WizardScene<MyWizardContext>(
         ctx.reply('Paljon pistetään?', {
             reply_markup: {
                 input_field_placeholder: '750',
-                // force the user to reply to the bot
                 force_reply: true,
             },
         });
-
         return ctx.wizard.next();
     },
     async (ctx) => {
-        // guard to check if the chat is valid
         if (!('text' in ctx.message)) {
             ctx.reply('Bettaas nyt järkevästi');
             return ctx.scene.reenter();
         }
+
         const bet = parseFloat(ctx.message.text);
         if (isNaN(bet)) {
             ctx.reply('Syötä betti muodossa 100 (ilman desimaaleja)', {
@@ -110,9 +104,7 @@ export const betWizard = new Scenes.WizardScene<MyWizardContext>(
             });
             return ctx.scene.reenter();
         } else {
-            // get current bet from db and check if it's below the new one
             const currentBet = await db.getBet(ctx.message.from.id);
-
             if (currentBet && currentBet > bet) {
                 ctx.reply(`Et voi betata vähemmän kuin ${currentBet} 😡`, {
                     reply_markup: {
@@ -122,18 +114,23 @@ export const betWizard = new Scenes.WizardScene<MyWizardContext>(
                 return ctx.scene.reenter();
             }
 
-            await db.setBet(
+            const result = await db.setBet(
                 ctx.message.from.id,
                 ctx.message.from.first_name,
                 ctx.message.from.last_name,
                 bet,
             );
-            await ctx.replyWithPhoto(
-                { source: 'heinis.jpg' },
-                {
-                    caption: `💥 Erinomainen betti ${ctx.message.from.first_name}! 💥`,
-                },
-            );
+            if (result.success) {
+                await ctx.replyWithPhoto(
+                    { source: 'heinis.jpg' },
+                    {
+                        caption: `💥 Erinomainen betti ${ctx.message.from.first_name}! 💥`,
+                    },
+                );
+            } else {
+                ctx.reply('Jokin meni pieleen, yritä uudelleen.');
+                return ctx.scene.reenter();
+            }
         }
         return ctx.scene.leave();
     },
@@ -145,30 +142,30 @@ export const nicknameWizard = new Scenes.WizardScene<MyWizardContext>(
         ctx.reply('Anna uusi lempinimi', {
             reply_markup: {
                 input_field_placeholder: 'Hessu',
-                // force the user to reply to the bot
                 force_reply: true,
             },
         });
-
         return ctx.wizard.next();
     },
     async (ctx) => {
-        // guard to check if the chat is valid
         if (!('text' in ctx.message)) {
             ctx.reply('Vastaa nyt järkevästi');
             return ctx.scene.reenter();
         }
 
         const nickname = ctx.message.text;
-
         const result = await db.setNickname(ctx.message.from.id, nickname);
 
-        await ctx.reply(`Lempinimesi on nyt ${nickname}`, {
-            reply_markup: {
-                remove_keyboard: true,
-            },
-        });
-
+        if (result.success) {
+            ctx.reply(`Lempinimesi on nyt ${nickname}`, {
+                reply_markup: {
+                    remove_keyboard: true,
+                },
+            });
+        } else {
+            ctx.reply('Jokin meni pieleen, yritä uudelleen.');
+            return ctx.scene.reenter();
+        }
         return ctx.scene.leave();
     },
 );
